Fix regression for checking validity of login
Unfortunately a regression was introduced which means that sessions are not automatically re-validated when the session expires (according to the expiry time provided by Google). This fixes that and also provides the `login_hint` e-mail address to avoid the account chooser on re-auth.

Note that this might break applications with AJAX endpoints as calls within JS will not follow the redirect correctly. The workflow project has a way around this - speak to @steppenwells if interested. Alternatively you can relax the requirement to re-authenticate by setting the `enforceValidity` parameter in your `GoogleAuthConfig` to `false`.