Tweak GoogleGroupChecker API to allow checking multiple groups
This follows on from https://github.com/guardian/play-googleauth/pull/20 -
we just need to check membership of multiple groups now, so it's better to
just return the set of returned groups, rather than do immediate filtering
which loses us that data.

AFAIK Membership is currently the only consumer of the GoogleGroupChecker
class, so I don't think the change will break any other clients.

I've also made it clearer that this is actually a blocking call, by
returning a Future and using `blocking` to help the execution context
know it may need to allocate a thread.