Add some Play 2.4-friendly overloads
If you're using Play 2.4 with compile-time DI, it's a little tricky to
use play-googleauth in its current state because it depends on
extracting the WS client from the currently running app. This means:

* you need to import `Play.current`, which is frowned upon in Play 2.4
* you have to add the WS components to the application's injector

Added some overloads that allow you to pass a WS client (or rather a
`WSAPI`, to be precise) explicitly, which avoids this mess.