Add a 'requireGroup' filter that checks Google Group membership
I've included an updated example to show how to use the filter, this is
the core bit:

```
type GoogleAuthRequest[A] = AuthenticatedRequest[A, googleauth.UserIdentity]

def authenticatedAndInGroup = (AuthAction andThen requireGroup[GoogleAuthRequest](includedGroups = Set("membership.dev@guardian.co.uk"))) {
  request => Ok(views.html.authenticated(request))
}
```

Note that I _haven't_ configured an example `GoogleServiceAccount`, as
those credentials would allow anyone to read Guardian Google Group
membership!

This code is taken from membership-frontend, where we quite often have a
request type which isn't just a
`AuthenticatedRequest[A, googleauth.UserIdentity]` (for instance, it might
contain Membership data!)- so I've had to do a bit of refactoring
(introducing `UserIdentifier`) so that the filter can find out the user's
identity using the `userIdentity()` method, not by taking it from `Request`
object (which may not be a `AuthenticatedRequest`, or directly contain a
`googleauth.UserIdentity`).