Make it much harder to accidentally fail to provide an email domain
The domain parameter is used to restrict logins to a specific email
domain, rather than allowing any google mail account. It's very easy
to forget to include this option or to have it be accidentally removed
without anyone noticing.

This change removes the GoogleAuthConfig constructor and provides two
different methods for creating instances. The default method (`apply`)
now explicitly requires a domain. An alternative is provided that does
not require a domain and it is named to make it very clear that this
will not include domain protection.

While this is a "general-purpose" library in practice the only users
are at the Guardian. Our applications all need to enforce a Guardian
email check and this makes writing correct code the path of least
resistance.

I think `5.3` should be the next version number but since this is a
minor breaking change so I'm open to other suggestions.