Overhaul to support Play 2.6
Play 2.6 requires many things to have a `BodyParser` parameter. This doesn't fit nicely with the existing model, so I've split the existing function into two parts.

`AuthAction` is now a separate object which would normally be passed into the Controller, while `LoginSupport` is only expected to be used by a controller specifically handling the details of Login.