Bring play-googleauth library to the top level, push example down
This is bit of style-based change - I've always found it a bit irksome
that the code for the actual library is pushed down one directory layer,
and it can be confusing for newcomers to the code who run `sbt` in the
top level directory and find no build! A root-level `build.sbt` file is
preferable to me.

This change brings the library code up to the top level, and uses the
`scripted` plugin to allow automatic compilation of the `example` project
(which is now located at `src/sbt-test/example/webapp`). The `scripted`
plugin is intended for plugin development, but it can also compile/run
full sbt projects if you want it to:

https://www.scala-sbt.org/1.x/docs/Testing-sbt-plugins.html#step+4%3A+write+a+script
https://stackoverflow.com/q/50202273/438886

I've also added a Travis CI file, which will run the `scripted` command and
so check that the example project compiles. I'm using Travis CI rather than
TeamCity as this is a public open-source project, and it's good for
external devs to see the CI results.

Note that I think the historical intention of the `example` project is to
provide a completely self-contained sbt build of an example project, as
reference for users of the library. This means that just making the whole
repo a multi-module build would be losing that standalone-example goal.

I've also updated the README to point to the new location of the example
project.