Support both Play version v2.6 & v2.7 (in a single release!)
The Maven coordinates for this project's artifacts have changed with this
commit:

*    group name: from `com.gu` to `com.gu.play-googleauth`
* artifact name: from `play-googleauth` to `play-v26` & `play-v27`

...so the final coordinate goes from:

```
"com.gu" %% "play-googleauth" % "[maven version number]"
```

to

```
"com.gu.play-googleauth" %% "play-v26" % "[maven version number]"
```

or

```
"com.gu.play-googleauth" %% "play-v27" % "[maven version number]"
```

As the coordinates have changed so significantly, I've also bumped the
version number to 1.0.0.

The `play-googleauth` code for Play v2.6 & v2.7 is thankfully identical,
so I've placed the reference source code in the `play-v26` folder, and
created a symlink to the `src` in that folder from within the `play-v27`
folder. This means you can view the code against both versions of Play in
your IDE, and see what compiles and what doesn't as you change it.

This commit also updates the `sbt-sonatype` plugin - apparently this
version is much faster to upload to Sonatype!

https://users.scala-lang.org/t/10x-faster-release-with-sbt-sonatype-3-6/5046
https://twitter.com/taroleo/status/1169777599360331776