Adopt GHA Scala Library Release Workflow
This is the 3rd library to adopt https://github.com/guardian/gha-scala-library-release-workflow,
a new automated library release process that allows releases to be
published to Maven Central with the click of a button.

The changes in here are similar to the ones in
https://github.com/guardian/play-secret-rotation/pull/416, with the addition
that sbt settings that provide the pom metadata required by Maven Central
( https://central.sonatype.org/publish/requirements/#project-name-description-and-url )
can be simplified thanks to the functionality provided by
https://github.com/xerial/sbt-sonatype/pull/58 - setting
`sonatypeProjectHosting` takes care of a lot the fields!