Remove old comment about dependency
This comment was originally introduced in
https://github.com/guardian/play-googleauth/pull/83, I think to indicate
that the dependency had been added effectively as an override, i.e. to
increase the version of a transitive dependency and avoid a snyk
vulnerability.

However, we actually import from this dependency and
sbt-explicit-dependencies reports it as an undeclared dependency when we
remove it: so we should keep it declared as a dependency.

(Also, the transitive version is now new enough that the specific snyk
version is no longer relevant.)